+++ app/ante/testutils/testutils.go
@@ -20,7 +20,6 @@ import (
 	"github.com/evmos/evmos/v20/testutil/integration/evmos/keyring"
 	"github.com/evmos/evmos/v20/testutil/integration/evmos/network"
 	"github.com/evmos/evmos/v20/types"
-	"github.com/evmos/evmos/v20/x/evm/config"
 	evmtypes "github.com/evmos/evmos/v20/x/evm/types"
 	feemarkettypes "github.com/evmos/evmos/v20/x/feemarket/types"
 )
@@ -90,18 +89,27 @@ func (suite *AnteTestSuite) SetupTest() {
 
 	suite.Require().NotNil(suite.network.App.AppCodec())
 
-	chainConfig := config.DefaultChainConfig(suite.network.GetChainID())
+	chainConfig := evmtypes.DefaultChainConfig(suite.network.GetChainID())
 	if !suite.enableLondonHF {
 		maxInt := sdkmath.NewInt(math.MaxInt64)
-		chainConfig.LondonBlock = maxInt.BigInt()
-		chainConfig.ArrowGlacierBlock = maxInt.BigInt()
-		chainConfig.GrayGlacierBlock = maxInt.BigInt()
-		chainConfig.MergeNetsplitBlock = maxInt.BigInt()
-		chainConfig.ShanghaiBlock = maxInt.BigInt()
-		chainConfig.CancunBlock = maxInt.BigInt()
+		chainConfig.LondonBlock = &maxInt
+		chainConfig.ArrowGlacierBlock = &maxInt
+		chainConfig.GrayGlacierBlock = &maxInt
+		chainConfig.MergeNetsplitBlock = &maxInt
+		chainConfig.ShanghaiBlock = &maxInt
+		chainConfig.CancunBlock = &maxInt
 	}
-	err := config.NewEVMConfigurator().
+
+	// get the denom and decimals set when initialized the chain
+	// to set them again
+	// when resetting the chain config
+	denom := evmtypes.GetEVMCoinDenom()       //nolint:staticcheck
+	decimals := evmtypes.GetEVMCoinDecimals() //nolint:staticcheck
+	configurator := evmtypes.NewEVMConfigurator()
+	configurator.ResetTestChainConfig()
+	err := configurator.
 		WithChainConfig(chainConfig).
+		WithEVMCoinInfo(denom, decimals).
 		Configure()
 	suite.Require().NoError(err)