+++ x/evm/keeper/state_transition_test.go
@@ -26,7 +26,6 @@ import (
 	"github.com/evmos/evmos/v20/testutil/integration/evmos/network"
 	"github.com/evmos/evmos/v20/testutil/integration/evmos/utils"
 	utiltx "github.com/evmos/evmos/v20/testutil/tx"
-	"github.com/evmos/evmos/v20/x/evm/config"
 	"github.com/evmos/evmos/v20/x/evm/keeper"
 	"github.com/evmos/evmos/v20/x/evm/types"
 	feemarkettypes "github.com/evmos/evmos/v20/x/feemarket/types"
@@ -256,10 +255,10 @@ func (suite *KeeperTestSuite) TestGetEthIntrinsicGas() {
 
 	for _, tc := range testCases {
 		suite.Run(fmt.Sprintf("Case %s", tc.name), func() {
-			ethCfg := config.GetChainConfig()
+			ethCfg := types.GetChainConfig()
 			ethCfg.HomesteadBlock = big.NewInt(2)
 			ethCfg.IstanbulBlock = big.NewInt(3)
-			signer := gethtypes.LatestSignerForChainID(config.GetChainConfig().ChainID)
+			signer := gethtypes.LatestSignerForChainID(types.GetChainConfig().ChainID)
 
 			ctx := suite.network.GetContext().WithBlockHeight(tc.height)
 
@@ -361,7 +360,7 @@ func (suite *KeeperTestSuite) TestRefundGas() {
 	// for refund to work
 	// NOTE: everything should happen within the same block for
 	// feecollector account to remain funded
-	baseDenom := config.GetEVMCoinDenom()
+	baseDenom := types.GetEVMCoinDenom()
 
 	coins := sdk.NewCoins(sdk.NewCoin(
 		baseDenom,
@@ -544,13 +543,13 @@ func (suite *KeeperTestSuite) TestEVMConfig() {
 	suite.Require().Equal(types.DefaultParams(), cfg.Params)
 	// london hardfork is enabled by default
 	suite.Require().Equal(big.NewInt(0), cfg.BaseFee)
-	suite.Require().Equal(config.GetChainConfig(), cfg.ChainConfig)
+	suite.Require().Equal(types.GetChainConfig(), cfg.ChainConfig)
 
 	validators := suite.network.GetValidators()
 	proposerHextAddress := utils.ValidatorConsAddressToHex(validators[0].OperatorAddress)
 	suite.Require().Equal(proposerHextAddress, cfg.CoinBase)
 
-	networkConfig := config.GetChainConfig()
+	networkConfig := types.GetChainConfig()
 	suite.Require().Equal(networkConfig, cfg.ChainConfig)
 }